pyproject.toml
[project]
name = "bdideepagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.3.1",
    "dotenv>=0.9.9",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "tavily-python>=0.7.17",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
